# PERF: Rely on C-level str conversions in loadtxt for up to 2x speedup