@seberg 2021-09-22
Mainly bringing it up in case it interests you, @anntzer.  But part of the reason the momentum has stalled a bit here is that @rossbar and I have been looking to pushing forward [`npreadtext`](https://github.com/BIDS-numpy/npreadtext/) with the goal of replacing `np.loadtxt` by moving it to C: https://mail.python.org/archives/list/numpy-discussion@python.org/thread/X4AU2DUDDNA44HTEFDQXJLC24E6MDEW3/

That does not have to stand in the way here though.   But, it would give us a good speed-up and additionally allows supporting new features, such as `quote='"'` (and other `csv.Dialect` features in the future), or even user provided C-parsers.