@anntzer 2021-09-22
That sounds great; I guess it depends on the timescale over which you think npreadtext will make it into numpy.  From a quick test npreadtext is faster than loadtxt even with the improvements here, but I am slightly worried that including a large chunk of C may take a while to review (wearing my matplotlib dev hat here), whereas the PR here may be faster to review (although it also involves some tricks).

So if you think npreadtext can be merged relatively quickly (wrt. numpy's release schedule), I am fine with closing this PR and its followup; otherwise, perhaps it can still go in as a temporary stopgap improvement.