@seberg 2022-01-16
Closing, as superseded by gh-20580, I don't think it is helpful to keep this open, unless the other PR gets rejected (which at this point seems a very long shot, I think it is finished and good).